Extract proof functions to proof_utils.rs, refactoring tests
