+++ libvcx/src/lib.rs
@@ -49,6 +49,7 @@ pub mod object_cache;
 pub mod disclosed_proof;
 
 pub mod v3;
+mod proof_utils;
 
 #[allow(unused_imports)]
 #[allow(dead_code)]
